api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
google-cloud-logging
google-cloud-aiplatform
google-cloud-discoveryengine
google-generativeai
pydantic
